// Reactivities/reactivities.ui/src/app/models/profile.ts
import { User } from "./user";

export interface Profile {
    userName: string;
    displayName: string;
    image?: string;
    bio?: string;
    followersCount: number;
    followingCount: number;
    following: boolean;
    photos: Photo[];
}          

export class Profile implements Profile {
    constructor(user: User){
        this.userName = user.username;
        this.displayName = user.displayName;
        this.image = user.image;
    }
}

export interface Photo {
    id: string;
    url: string;
    isMain: boolean;
}

export interface UserActivity {
    id: string,
    title: string,
    date: Date,
    hostUsername: string;
    category: string;
}